+++ perf/SourceGeneratorPerf/SourceGeneratorPerf/CompilationHelper.cs
@@ -27,7 +27,7 @@ public static class CompilationHelper
             MetadataReference.CreateFromFile(typeof(MessagePack.MessagePackObjectAttribute).Assembly.Location), // MessagePack.Annotations
             MetadataReference.CreateFromFile(typeof(MagicOnion.IService<>).Assembly.Location), // MagicOnion.Abstractions
             MetadataReference.CreateFromFile(typeof(MagicOnion.Client.MagicOnionClient).Assembly.Location), // MagicOnion.Client
-            MetadataReference.CreateFromFile(typeof(MagicOnion.GrpcMethodHelper).Assembly.Location), // MagicOnion.Shared
+            MetadataReference.CreateFromFile(typeof(MagicOnion.Internal.GrpcMethodHelper).Assembly.Location), // MagicOnion.Shared
         };
         var compilationOptions = new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary);
