+++ tests/MagicOnion.Client.SourceGenerator.Tests/CompilationHelper.cs
@@ -28,7 +28,7 @@ public static class CompilationHelper
             MetadataReference.CreateFromFile(typeof(MessagePack.MessagePackObjectAttribute).Assembly.Location), // MessagePack.Annotations
             MetadataReference.CreateFromFile(typeof(MagicOnion.IService<>).Assembly.Location), // MagicOnion.Abstractions
             MetadataReference.CreateFromFile(typeof(MagicOnion.Client.MagicOnionClient).Assembly.Location), // MagicOnion.Client
-            MetadataReference.CreateFromFile(typeof(MagicOnion.GrpcMethodHelper).Assembly.Location), // MagicOnion.Shared
+            MetadataReference.CreateFromFile(typeof(MagicOnion.Internal.GrpcMethodHelper).Assembly.Location), // MagicOnion.Shared
             MetadataReference.CreateFromFile(typeof(IGreeterService).Assembly.Location), // SampleServiceDefinitions
         };
         var compilationOptions = new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary);
